backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu


qdrant-client
sentence-transformers
pillow
pandas
numpy
requests
python-dotenv
fastapi
uvicorn
python-multipart
torch==2.3.1+cpu
tqdm
